update : price를 int -> long으로 변경

+++ src/main/java/com/limitedgoods/limitedgoods/order/dto/OrderPaymentInfo.java
@@ -4,7 +4,7 @@ import com.limitedgoods.limitedgoods.order.entity.OrderStatus;
 
 public record OrderPaymentInfo(
         Long orderId,
-        int totalPrice,
+        long totalPrice,
         OrderStatus orderStatus
 ) {
 }
